Use the Android Property Animation framework to animate the UI to indicate a file is being downloaded. This replaces the usage of MotionLayout in the main fragment.
Add animation to download detail fragment using MotionLayout. This is used to add a transition for when the views are first displayed in the fragment_download_details.xml. In addition use a transition in MotionLayout to direct the users attention to the status of a failed download.

Add placeholder text so the preview of the fragment_download_details.xml better represents the user's experience.